typescriptify
